requirements.txt
Flask==3.0.0
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==3.0.0
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
requests==2.31.0